Refactoring - part 1 (#4)
As part of adding the unit tests, a lot of refactoring and code cleanup is taking place. Here we mainly try to separate the csv_lib functionalities and the hdf5_lib functionalities in a better way; cleanup the database class; and come up with a good small example to test in the unit tests.